refactor: extracted MainNav from Container
